Move reference manager setup code to form package
